requirements.txt
speedtest-cli>=2.1.3
PyYAML>=6.0.1
mysql-connector-python>=8.3.0